Add dependency: log
This is the same `log` (and version) module that's being used in the irc
adapter. The reason for adding this is so that we can get rid of
`console.log`'s and instead make use of a more sane library.

When writing tests some of these `console.log`'s will output information
which would require one to mock out `console.log` and then restore it
before each test is done running. Very tedious.